ranking/requirements.txt
pymongo==3.12.0
boto3==1.26.9
python-dotenv==0.21.0
